backend/requirements.txt
# Flask and Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-openai==0.0.2
langchain-together==0.0.1
langchain-pinecone==0.0.1

# OpenAI and AI Models
openai==1.6.1
anthropic==0.8.1

# Vector Databases
faiss-cpu==1.7.4
pinecone-client==3.0.0

# Document Processing
unstructured==0.12.4
unstructured-client==0.18.0
pdf2image==1.16.3
pdfplumber==0.10.3
pypdf==3.17.4

# PDF and Document Utilities
python-docx==1.1.0
docx2pdf==0.1.8

# Data Processing
pandas==2.1.4
numpy==1.24.3
beautifulsoup4==4.12.2

# Evaluation
ragas==0.1.1
datasets==2.16.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
googlesearch-python==1.2.3
selenium==4.16.0

# Image Processing
pillow==10.1.0

